+++ src/components/modal/TModalBox.vue
@@ -246,7 +246,7 @@ const combinedStyle = computed<CSSProperties>(() => {
   left: 0;
   width: 100%;
   height: 100%;
-  background-color: rgba(0, 0, 0, 0.5);
+  background-color: var(--t-color-backdrop);
   backdrop-filter: blur(3px);
   z-index: 3990;
   pointer-events: auto;